+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/config/ClientConfig.kt
@@ -17,6 +17,7 @@ class ClientConfig(
         lateinit var disableParrotNarrator: ModConfigSpec.BooleanValue
         lateinit var disableCringeLangChange: ModConfigSpec.BooleanValue
         lateinit var supportSlimeRenderDebug: ModConfigSpec.BooleanValue
+        lateinit var itemMagnetBlacklistTooltip: ModConfigSpec.BooleanValue
     }
 
     init {
@@ -43,6 +44,10 @@ class ClientConfig(
         supportSlimeRenderDebug = builder
             .comment("Enable to render the base Slime model for Support Slime entities.")
             .define("supportSlimeRenderDebug", false)
+
+        itemMagnetBlacklistTooltip = builder
+            .comment("Show that an item is in the Item Magnet's blacklist in its tooltip.")
+            .define("itemMagnetBlacklistTooltip", true)
 
         builder.pop()
     }