+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/datagen/ModLanguageProvider.kt
@@ -156,6 +156,7 @@ class ModLanguageProvider(
         const val GMO_LINE_ONE = "tooltip.geneticsresequenced.gmo_recipe.line1"
         const val GMO_LINE_TWO = "tooltip.geneticsresequenced.gmo_recipe.line2"
         const val GMO_LINE_THREE = "tooltip.geneticsresequenced.gmo_recipe.line3"
+        const val ITEM_MAGNET_BLACKLIST = "tooltip.geneticsresequenced.item_magnet_blacklist"
     }
 
     object Commands {
@@ -576,6 +577,7 @@ class ModLanguageProvider(
         )
         add(Tooltips.GMO_LINE_TWO, "Otherwise, it always produces a %s Gene.")
         add(Tooltips.GMO_LINE_THREE, "Overclockers lower your chance, but Chorus Fruits increase it!")
+        add(Tooltips.ITEM_MAGNET_BLACKLIST, "Blacklisted from Item Magnet Gene")
 
         add(Commands.LIST_ALL_GENES, "Gene List:")
         add(Commands.NO_GENES, "No Genes found!")